refactor: move navigation constants from constants.ts to navigation.ts and update imports
