+++ src/utils/routes.ts
@@ -1,7 +1,7 @@
 /* eslint-disable import/prefer-default-export */
 import { NextRouter } from 'next/router';
 
-import { AUTH_ROUTES } from './constants';
+import { AUTH_ROUTES } from './navigation';
 
 /**
  * Check if the current route is an authentication page
